+++ core/src/elements/ons-tabbar/index.spec.js
@@ -4,8 +4,8 @@ describe('OnsTabbarElement', () => {
   let element, template, template2;
 
   beforeEach(done => {
-    template = ons._util.createElement('<ons-template id="hoge"><ons-page>hogehoge</ons-page></ons-template>');
-    template2 = ons._util.createElement('<ons-template id="fuga"><ons-page>fugafuga</ons-page></ons-template>');
+    template = ons._util.createElement('<template id="hoge"><ons-page>hogehoge</ons-page></template>');
+    template2 = ons._util.createElement('<template id="fuga"><ons-page>fugafuga</ons-page></template>');
     element = new ons.TabbarElement();
     document.body.appendChild(template);
     document.body.appendChild(template2);
@@ -165,8 +165,8 @@ describe('OnsTabbarElement', () => {
       const div = document.createElement('div');
       document.body.appendChild(div);
       div.innerHTML = `
-        <ons-template id="page1"></ons-template>
-        <ons-template id="page2"></ons-template>
+        <template id="page1"></template>
+        <template id="page2"></template>
         <ons-tabbar id="myTabbar">
           <ons-tab id="tab1" page="page1"></ons-tab>
           <ons-tab id="tab2" page="page2"></ons-tab>
@@ -193,7 +193,7 @@ describe('OnsTabbarElement', () => {
     let template, element;
 
     beforeEach(() => {
-      template = ons._util.createElement('<ons-template id="page1"><ons-page>Page1</ons-page></ons-template>');
+      template = ons._util.createElement('<template id="page1"><ons-page>Page1</ons-page></template>');
       element = ons._util.createElement(`
         <ons-tabbar>
           <ons-tab label="Page 1" page="page1" no-reload></ons-tab>
@@ -240,7 +240,7 @@ describe('OnsTabbarElement', () => {
     let template, element;
 
     beforeEach(() => {
-      template = ons._util.createElement('<ons-template id="page1"><ons-page>Page1</ons-page></ons-template>');
+      template = ons._util.createElement('<template id="page1"><ons-page>Page1</ons-page></template>');
       element = ons._util.createElement(`
         <ons-tabbar>
           <ons-tab label="Page 1" page="page1"></ons-tab>