+++ core/src/elements/ons-toast/ascend-animator.js
@@ -18,6 +18,7 @@ limitations under the License.
 
 import util from '../../ons/util';
 import animit from '../../ons/animit';
+import platform from '../../ons/platform';
 import ToastAnimator from './animator';
 
 /**
@@ -29,7 +30,7 @@ export default class AscendToastAnimator extends ToastAnimator {
     super({ timing, delay, duration });
 
     this.messageDelay = this.duration * 0.4 + this.delay; // Delay message opacity change
-    this.ascension = undefined; // Calculated during the first animation
+    this.ascension = platform.isAndroid() ? 48 : 64; // Toasts are always 1 line
   }
 
   /**
@@ -37,13 +38,8 @@ export default class AscendToastAnimator extends ToastAnimator {
    * @param {Function} callback
    */
   show(toast, callback) {
-    const siblingPage = this._findSiblingControlComponent(toast);
     toast = toast._toast;
-
-    if (!this.ascension) {
-      const cs = window.getComputedStyle(toast);
-      this.ascension = parseInt(cs.getPropertyValue('height'), 10) + parseInt(cs.getPropertyValue('margin-bottom'), 10);
-    }
+    util.globals.fabOffset = this.ascension;
 
     animit.runAll(
       animit(toast)
@@ -64,13 +60,10 @@ export default class AscendToastAnimator extends ToastAnimator {
           done();
         }),
 
-      animit(siblingPage) // Cannot save/restore style
-        .queue({
-          bottom: '0'
-        })
+      animit(this._getFabs())
         .wait(this.delay)
         .queue({
-          bottom: `${this.ascension}px`
+          transform: `translate3d(0, -${this.ascension}px, 0) scale(1)`
         }, {
           duration: this.duration,
           timing: this.timing
@@ -97,13 +90,8 @@ export default class AscendToastAnimator extends ToastAnimator {
    * @param {Function} callback
    */
   hide(toast, callback) {
-    const siblingPage = this._findSiblingControlComponent(toast);
     toast = toast._toast;
-
-    if (!this.ascension) {
-      const cs = window.getComputedStyle(toast);
-      this.ascension = parseInt(cs.getPropertyValue('height'), 10) + parseInt(cs.getPropertyValue('margin-bottom'), 10);
-    }
+    util.globals.fabOffset = 0;
 
     animit.runAll(
       animit(toast)
@@ -124,13 +112,10 @@ export default class AscendToastAnimator extends ToastAnimator {
           done();
         }),
 
-      animit(siblingPage) // Cannot save/restore style
-        .queue({
-          bottom: `${this.ascension}px`
-        })
+      animit(this._getFabs())
         .wait(this.delay)
         .queue({
-          bottom: '0'
+          transform: 'translate3d(0, 0, 0) scale(1)'
         }, {
           duration: this.duration,
           timing: this.timing
@@ -152,8 +137,7 @@ export default class AscendToastAnimator extends ToastAnimator {
     );
   }
 
-  _findSiblingControlComponent(toast) {
-    const query = e => /(ons-page|ons-navigator|ons-tabbar|ons-splitter)/i.test(e.tagName);
-    return util.findChild(toast.parentNode, query) || util.findChild(document.body.children[0], query);
+  _getFabs() {
+    return util.arrayFrom(document.querySelectorAll('ons-fab[position~=bottom], ons-speed-dial[position~=bottom]')).filter(fab => fab.visible);
   }
 }