+++ CHANGELOG.md
@@ -7,14 +7,21 @@ dev
 
 ### New Features
 
+ * core: Separated files are now cached after the first request. This improves performance when pushing pages.
+ * core: Added `ons.preload` method to manually cache templates. This improves performance when pushing pages.
+ * core: Support for `HTMLTemplateElement` (`<template>`). The former `<ons-template>` is still supported for backward compatibility.
+ * core: Added page life-cycle hooks (similar to life-cycle events). `<template>` elements allow `<script>` tags in their content, which is useful for initialising pages. Therefore, hooks like `pageElement.onInit`, `pageElement.onShow`, etc. have been added.
 
 ### Bug Fixes
 
 
 ### Misc
 
+ * css-components: Refined the appearance of `checkbox` on iOS.
+ * css-components: Refined the appearance of `radio-button` on iOS.
  * css-components: Refined the appearance of `button` on Android.
  * core: Updated `font-awesome` icon library.
+ * ons-toast: Improved 'ascend' animator performance.
 
 ### BREAKING CHANGES